refactor: consolidate modal state management by creating a unified atoms file for CreateKiwi and DetailKiwi modals
