+++ apps/web/src/routes/my-kiwis/-modals/detail-kiwi/atoms.ts
@@ -1,9 +1,4 @@
 import { atom } from "@bookiwi/jotai";
 import { MyKiwi } from "@bookiwi/supabase/types";
 
-export const kiwiDetailModalOpenAtom = atom<boolean>(false);
 export const selectedKiwiAtom = atom<MyKiwi | null>(null);
-export const openKiwiDetailModalAtom = atom(null, (get, set, kiwi: MyKiwi) => {
-  set(kiwiDetailModalOpenAtom, true);
-  set(selectedKiwiAtom, kiwi);
-});
